Fixing some functionally issues
I attempted to do the enum method for a while and after contacting virtual services they thought that given that it is just 3 attributes that it might be easier to have it as if statements with strings. Did it to be able to progress through the walkthrough as I was getting a lot of errors with enums.